Log Widgets moved to new file. Time Stamp logic updated for log widgets
